removed predicate resolver
