--- Fb2.Document/PredicateResolver/PredicateResolver.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using Fb2.Document.Models.Base;
-
-namespace Fb2.Document.Resolver;
-
-internal static class PredicateResolver
-{
-    private static Func<Fb2Node, bool> GetAbstractClassPredicate(Type targetType)
-        => element => element.GetType().IsSubclassOf(targetType);
-
-    private static Func<Fb2Node, bool> GetClassPredicate(Type targetType)
-        => element => element.GetType().Equals(targetType);
-
-    internal static Func<Fb2Node, bool> GetPredicate<T>() where T : Fb2Node
-    {
-        var targetType = typeof(T);
-
-        if (targetType.IsAbstract)
-            return GetAbstractClassPredicate(targetType);
-
-        return GetClassPredicate(targetType);
-    }
-}
